# ENH: Enable SVE detection for Highway VQSort