@Mousius 2023-11-24
@rdevulap don't panic; this was just so easy to do that I wanted to see if it would work 😸 I'll wait for your patches to be in and rebase on top of them.

@seiko2plus SVE features! Woo!