@seiko2plus 2023-11-29
>  this was just so easy to do that I wanted to see if it would work 😸 I'll wait for your patches to be in and rebase on top of them.

This pr doesn't change any C/C++ code so it should be fine to merge it.