@Mousius 2023-11-29
> > this was just so easy to do that I wanted to see if it would work 😸 I'll wait for your patches to be in and rebase on top of them.
> 
> This pr doesn't change any C/C++ code so it should be fine to merge it.

This will conflict with https://github.com/numpy/numpy/pull/25045/files#diff-1df2f2cfde074ed7de97ebae8a7ef48ad000241256ea3f480f81b2598d03b4f8R784-R788